@@ -1,20 +1,19 @@
 ﻿using System;
 using FluentValidation;
 
-namespace API.Contracts.Requests.v1
-{
-    public record UpdateCommentRequestV1
-    {
-        public Guid CommentId { get; set; }
-        public string Text { get; set; }
-    }
+namespace API.Contracts.Requests.v1;
 
-    public class UpdateCommentRequestV1Validation : AbstractValidator<UpdateCommentRequestV1>
+public record UpdateCommentRequestV1
+{
+    public Guid CommentId { get; set; }
+    public string Text { get; set; }
+}
+
+public class UpdateCommentRequestV1Validation : AbstractValidator<UpdateCommentRequestV1>
+{
+    public UpdateCommentRequestV1Validation()
     {
-        public UpdateCommentRequestV1Validation()
-        {
-            RuleFor(m => m.CommentId).NotEmpty();
-            RuleFor(m => m.Text).NotEmpty();
-        }
+        RuleFor(m => m.CommentId).NotEmpty();
+        RuleFor(m => m.Text).NotEmpty();
     }
 }
