@@ -1,10 +1,9 @@
 ﻿using API.Dtos;
 
-namespace API.Contracts.Responses.v1
+namespace API.Contracts.Responses.v1;
+
+public record AuthorizedResponseV1
 {
-    public record AuthorizedResponseV1
-    {
-        public string Token { get; set; }
-        public UserDto User { get; set; }
-    }
+    public string Token { get; set; }
+    public UserDto User { get; set; }
 }
