@@ -1,15 +1,14 @@
 ﻿using System.Text.Json;
 
-namespace API.Contracts.Responses.v1
-{
-    public class FailedResponseV1
-    {
-        public int Code { get; set; }
-        public string Message { get; set; }
+namespace API.Contracts.Responses.v1;
 
-        public override string ToString()
-        {
-            return JsonSerializer.Serialize(this);
-        }
+public class FailedResponseV1
+{
+    public int Code { get; set; }
+    public string Message { get; set; }
+
+    public override string ToString()
+    {
+        return JsonSerializer.Serialize(this);
     }
 }
