@@ -1,11 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 
-namespace API.Controllers.Base
-{
-    [ApiController]
-    [Route("api/v{version:apiVersion}/[controller]")]
-    public abstract class ApiController : ControllerBase
-    {
-        
-    }
-}
+namespace API.Controllers.Base;
+
+[ApiController]
+[Route("api/v{version:apiVersion}/[controller]")]
+public abstract class ApiController : ControllerBase {}
