@@ -1,11 +1,10 @@
 ﻿using System;
 
-namespace API.Dtos
+namespace API.Dtos;
+
+public record CommentDto
 {
-    public record CommentDto
-    {
-        public Guid Id { get; set; }
-        public string Text { get; set; }
-        public DateTime CreatedAt { get; set; }
-    }
+    public Guid Id { get; set; }
+    public string Text { get; set; }
+    public DateTime CreatedAt { get; set; }
 }
