@@ -1,12 +1,11 @@
 ﻿using System;
 
-namespace API.Dtos
+namespace API.Dtos;
+
+public class PostDto
 {
-    public class PostDto
-    {
-        public Guid Id { get; set; } 
-        public string Text { get; set; }
-        public UserDto Author { get; set; }
-        // TODO: add field published at
-    }
+    public Guid Id { get; set; } 
+    public string Text { get; set; }
+    public UserDto Author { get; set; }
+    // TODO: add field published at
 }
