@@ -1,26 +1,25 @@
 ﻿using System;
 
-namespace API
+namespace API;
+
+public class Env
 {
-    public class Env
+    public class TokenExpirationTime
     {
-        public class TokenExpirationTime
-        {
-            public static TimeSpan OneDay = TimeSpan.FromDays(1);
-            public static TimeSpan SevenDays = TimeSpan.FromDays(7);
-        }
+        public static TimeSpan OneDay = TimeSpan.FromDays(1);
+        public static TimeSpan SevenDays = TimeSpan.FromDays(7);
+    }
         
-        public class IdentityClaims
-        {
-            public const string ID = "id";
-            public const string ROLES = "role";
-        }
+    public class IdentityClaims
+    {
+        public const string ID = "id";
+        public const string ROLES = "role";
+    }
         
-        public class Roles
-        {
-            public const string USER = "User";
-            public const string ADMIN = "Admin";
-            public const string ALL = "User, Admin";
-        }
+    public class Roles
+    {
+        public const string USER = "User";
+        public const string ADMIN = "Admin";
+        public const string ALL = "User, Admin";
     }
 }
