@@ -1,12 +1,11 @@
 ﻿using System;
 
-namespace API.Exceptions
+namespace API.Exceptions;
+
+public class BadRequestRestException : Exception
 {
-    public class BadRequestRestException : Exception
+    public BadRequestRestException(string message)
+        : base(message)
     {
-        public BadRequestRestException(string message)
-            : base(message)
-        {
-        }
     }
 }
