@@ -5,42 +5,41 @@ using System.Text;
 using Microsoft.Extensions.Configuration;
 using Microsoft.IdentityModel.Tokens;
 
-namespace API.Helpers
+namespace API.Helpers;
+
+public class JwtHelper
 {
-    public class JwtHelper
+    private const string SecurityAlgorithm = SecurityAlgorithms.HmacSha512Signature;
+
+    private readonly SigningCredentials _signingCredentials;
+
+    public JwtHelper(IConfiguration configuration)
     {
-        private const string SecurityAlgorithm = SecurityAlgorithms.HmacSha512Signature;
+        string secretText = configuration["Jwt:Secret"];
+        byte[] secret = Encoding.ASCII.GetBytes(secretText);
 
-        private readonly SigningCredentials _signingCredentials;
+        var securityKey = new SymmetricSecurityKey(secret);
 
-        public JwtHelper(IConfiguration configuration)
+        _signingCredentials = new SigningCredentials(securityKey, SecurityAlgorithm);
+    }
+
+    private static readonly JwtSecurityTokenHandler JwtTokenHandler = new();
+
+    public string GenerateToken(string userId, string userRoles, TimeSpan duration)
+    {
+        var descriptor = new SecurityTokenDescriptor
         {
-            string secretText = configuration["Jwt:Secret"];
-            byte[] secret = Encoding.ASCII.GetBytes(secretText);
-
-            var securityKey = new SymmetricSecurityKey(secret);
-
-            _signingCredentials = new SigningCredentials(securityKey, SecurityAlgorithm);
-        }
-
-        private static readonly JwtSecurityTokenHandler JwtTokenHandler = new();
-
-        public string GenerateToken(string userId, string userRoles, TimeSpan duration)
-        {
-            var descriptor = new SecurityTokenDescriptor
+            Subject = new ClaimsIdentity(new Claim[]
             {
-                Subject = new ClaimsIdentity(new Claim[]
-                {
-                    new(Env.IdentityClaims.ID, userId, ClaimValueTypes.String),
-                    new(Env.IdentityClaims.ROLES, userRoles, ClaimValueTypes.String)
-                }),
-                Expires = DateTime.UtcNow.Add(duration),
-                SigningCredentials = _signingCredentials
-            };
+                new(Env.IdentityClaims.ID, userId, ClaimValueTypes.String),
+                new(Env.IdentityClaims.ROLES, userRoles, ClaimValueTypes.String)
+            }),
+            Expires = DateTime.UtcNow.Add(duration),
+            SigningCredentials = _signingCredentials
+        };
 
-            var token = JwtTokenHandler.CreateToken(descriptor);
+        var token = JwtTokenHandler.CreateToken(descriptor);
 
-            return JwtTokenHandler.WriteToken(token);
-        }
+        return JwtTokenHandler.WriteToken(token);
     }
 }
