@@ -10,123 +10,122 @@ using DAL.Entities;
 using Mapster;
 using Microsoft.EntityFrameworkCore;
 
-namespace API.Services.v1
+namespace API.Services.v1;
+
+public class CommentsServiceV1
 {
-    public class CommentsServiceV1
+    private readonly DataContext _context;
+
+    public CommentsServiceV1(DataContext context)
     {
-        private readonly DataContext _context;
+        _context = context;
+    }
 
-        public CommentsServiceV1(DataContext context)
+    public async Task<IEnumerable<CommentDto>> GetCommentsFromPostAsync(Guid postId)
+    {
+        Post post = await _context.Posts.Include(c => c.Comments).FirstOrDefaultAsync(m => m.Id == postId);
+
+        if (post is null)
         {
-            _context = context;
+            throw new BadRequestRestException("Post does not exists");
         }
 
-        public async Task<IEnumerable<CommentDto>> GetCommentsFromPostAsync(Guid postId)
+        return post.Comments.Adapt<IEnumerable<CommentDto>>();
+    }
+
+    public async Task<IEnumerable<CommentDto>> GetCommentsFromUserAsync(Guid userId)
+    {
+        User user = await _context.Users.Include(c => c.Comments).FirstOrDefaultAsync(m => m.Id == userId);
+
+        if (user is null)
         {
-            Post post = await _context.Posts.Include(c => c.Comments).FirstOrDefaultAsync(m => m.Id == postId);
-
-            if (post is null)
-            {
-                throw new BadRequestRestException("Post does not exists");
-            }
-
-            return post.Comments.Adapt<IEnumerable<CommentDto>>();
+            throw new BadRequestRestException("User does not exists");
         }
 
-        public async Task<IEnumerable<CommentDto>> GetCommentsFromUserAsync(Guid userId)
-        {
-            User user = await _context.Users.Include(c => c.Comments).FirstOrDefaultAsync(m => m.Id == userId);
-
-            if (user is null)
-            {
-                throw new BadRequestRestException("User does not exists");
-            }
-
-            return user.Comments.Adapt<IEnumerable<CommentDto>>();
-        }
+        return user.Comments.Adapt<IEnumerable<CommentDto>>();
+    }
         
-        public async Task<CommentDto> CreateCommentAsync(CreateCommentRequestV1 request, Guid userId)
+    public async Task<CommentDto> CreateCommentAsync(CreateCommentRequestV1 request, Guid userId)
+    {
+        // TODO: check if post published then create comment
+            
+        User user = await _context.Users.FindAsync(userId);
+
+        if (user is null)
         {
-            // TODO: check if post published then create comment
-            
-            User user = await _context.Users.FindAsync(userId);
-
-            if (user is null)
-            {
-                throw new BadRequestRestException("User does not exists");
-            }
-
-            await _context.Entry(user).Collection(c => c.Posts).LoadAsync();
-
-            var post = user.Posts.FirstOrDefault(m => m.Id == request.PostId);
-
-            if (post is null)
-            {
-                throw new BadRequestRestException("Post does not exists");
-            }
-            
-            var comment = new Comment
-            {
-                Text = request.Text,
-                Post = post,
-                Author = user
-            };
-
-            await _context.Comments.AddAsync(comment);
-
-            await _context.SaveChangesAsync();
-
-            return comment.Adapt<CommentDto>();
+            throw new BadRequestRestException("User does not exists");
         }
 
-        public async Task UpdateCommentAsync(UpdateCommentRequestV1 request, Guid userId)
+        await _context.Entry(user).Collection(c => c.Posts).LoadAsync();
+
+        var post = user.Posts.FirstOrDefault(m => m.Id == request.PostId);
+
+        if (post is null)
         {
-            // TODO: check if post published then create comment
+            throw new BadRequestRestException("Post does not exists");
+        }
             
-            User user = await _context.Users.FindAsync(userId);
+        var comment = new Comment
+        {
+            Text = request.Text,
+            Post = post,
+            Author = user
+        };
 
-            if (user is null)
-            {
-                throw new BadRequestRestException("User does not exists");
-            }
-            
-            await _context.Entry(user).Collection(c => c.Comments).LoadAsync();
-            
-            var comment = user.Comments.FirstOrDefault(m => m.Id == request.CommentId);
+        await _context.Comments.AddAsync(comment);
 
-            if (comment is null)
-            {
-                throw new BadRequestRestException("Comment does not exists");
-            }
+        await _context.SaveChangesAsync();
 
-            comment.Text = request.Text;
+        return comment.Adapt<CommentDto>();
+    }
+
+    public async Task UpdateCommentAsync(UpdateCommentRequestV1 request, Guid userId)
+    {
+        // TODO: check if post published then create comment
             
-            await _context.SaveChangesAsync();
+        User user = await _context.Users.FindAsync(userId);
+
+        if (user is null)
+        {
+            throw new BadRequestRestException("User does not exists");
+        }
+            
+        await _context.Entry(user).Collection(c => c.Comments).LoadAsync();
+            
+        var comment = user.Comments.FirstOrDefault(m => m.Id == request.CommentId);
+
+        if (comment is null)
+        {
+            throw new BadRequestRestException("Comment does not exists");
         }
 
-        public async Task DeleteCommentAsync(Guid commentId, Guid userId)
+        comment.Text = request.Text;
+            
+        await _context.SaveChangesAsync();
+    }
+
+    public async Task DeleteCommentAsync(Guid commentId, Guid userId)
+    {
+        // TODO: check if post published then create comment
+            
+        User user = await _context.Users.FindAsync(userId);
+
+        if (user is null)
         {
-            // TODO: check if post published then create comment
-            
-            User user = await _context.Users.FindAsync(userId);
-
-            if (user is null)
-            {
-                throw new BadRequestRestException("User does not exists");
-            }
-            
-            await _context.Entry(user).Collection(c => c.Comments).LoadAsync();
-            
-            var comment = user.Comments.FirstOrDefault(m => m.Id == commentId);
-
-            if (comment is null)
-            {
-                throw new BadRequestRestException("Comment does not exists");
-            }
-
-            _context.Remove(comment);
-            
-            await _context.SaveChangesAsync();
+            throw new BadRequestRestException("User does not exists");
         }
+            
+        await _context.Entry(user).Collection(c => c.Comments).LoadAsync();
+            
+        var comment = user.Comments.FirstOrDefault(m => m.Id == commentId);
+
+        if (comment is null)
+        {
+            throw new BadRequestRestException("Comment does not exists");
+        }
+
+        _context.Remove(comment);
+            
+        await _context.SaveChangesAsync();
     }
 }
