@@ -10,143 +10,142 @@ using DAL.Entities;
 using Mapster;
 using Microsoft.EntityFrameworkCore;
 
-namespace API.Services.v1
+namespace API.Services.v1;
+
+public class PostsServiceV1
 {
-    public class PostsServiceV1
+    private readonly DataContext _context;
+
+    public PostsServiceV1(DataContext context)
     {
-        private readonly DataContext _context;
+        _context = context;
+    }
 
-        public PostsServiceV1(DataContext context)
-        {
-            _context = context;
-        }
-
-        private int CalculateOffset(int page, int perPage) => page <= 0 ? 1 : page * perPage - perPage;
+    private int CalculateOffset(int page, int perPage) => page <= 0 ? 1 : page * perPage - perPage;
         
-        public IEnumerable<PostDto> GetPublishedPosts(GetPublishesRequestV1 request)
-        {
-            int skip = CalculateOffset(request.Page, request.PerPage);
-            int take = request.PerPage;
+    public IEnumerable<PostDto> GetPublishedPosts(GetPublishesRequestV1 request)
+    {
+        int skip = CalculateOffset(request.Page, request.PerPage);
+        int take = request.PerPage;
             
-            IEnumerable<Post> posts = _context.Publishes
-                .Include(publish => publish.Post)
-                .OrderByDescending(publish => publish.CreatedTimeStamp)
-                .Skip(skip)
-                .Take(take)
-                .Select(publish => publish.Post);
+        IEnumerable<Post> posts = _context.Publishes
+            .Include(publish => publish.Post)
+            .OrderByDescending(publish => publish.CreatedTimeStamp)
+            .Skip(skip)
+            .Take(take)
+            .Select(publish => publish.Post);
 
-            return posts.Adapt<IEnumerable<PostDto>>();
+        return posts.Adapt<IEnumerable<PostDto>>();
+    }
+
+    public async Task<PostDto> GetTopPostAsync(Guid userId)
+    {
+        User user = await _context.Users.FindAsync(userId);
+
+        if (user is null)
+        {
+            throw new BadRequestRestException("User does not exists");
         }
 
-        public async Task<PostDto> GetTopPostAsync(Guid userId)
+        var post = await _context.Posts.OrderBy(post => post.CreatedTimeStamp).LastOrDefaultAsync();
+
+        if (post is null)
         {
-            User user = await _context.Users.FindAsync(userId);
-
-            if (user is null)
-            {
-                throw new BadRequestRestException("User does not exists");
-            }
-
-            var post = await _context.Posts.OrderBy(post => post.CreatedTimeStamp).LastOrDefaultAsync();
-
-            if (post is null)
-            {
-                throw new BadRequestRestException("Post does not exists");
-            }
-            
-            return post.Adapt<PostDto>();
+            throw new BadRequestRestException("Post does not exists");
         }
+            
+        return post.Adapt<PostDto>();
+    }
         
-        public async Task<PostDto> CreateAsync(Guid userId, CreatePostRequestV1 request)
+    public async Task<PostDto> CreateAsync(Guid userId, CreatePostRequestV1 request)
+    {
+        User user = await _context.Users.FindAsync(userId);
+
+        if (user is null)
         {
-            User user = await _context.Users.FindAsync(userId);
+            throw new BadRequestRestException("User does not exists");
+        }
+            
+        var post = new Post
+        {
+            Text = request.Text,
+            Author = user
+        };
+            
+        await _context.Posts.AddAsync(post);
+            
+        await _context.SaveChangesAsync();
 
-            if (user is null)
-            {
-                throw new BadRequestRestException("User does not exists");
-            }
-            
-            var post = new Post
-            {
-                Text = request.Text,
-                Author = user
-            };
-            
-            await _context.Posts.AddAsync(post);
-            
-            await _context.SaveChangesAsync();
+        return post.Adapt<PostDto>();
+    }
 
-            return post.Adapt<PostDto>();
+    public async Task UpdateAsync(Guid userId, UpdatePostRequestV1 request)
+    {
+        User user = await _context.Users.FindAsync(userId);
+
+        if (user is null)
+        {
+            throw new BadRequestRestException("User does not exists");
         }
 
-        public async Task UpdateAsync(Guid userId, UpdatePostRequestV1 request)
+        await _context.Entry(user).Collection(u => u.Posts).LoadAsync();
+            
+        var post = user.Posts.FirstOrDefault(post => post.Id == request.Id);
+
+        if (post is null)
         {
-            User user = await _context.Users.FindAsync(userId);
-
-            if (user is null)
-            {
-                throw new BadRequestRestException("User does not exists");
-            }
-
-            await _context.Entry(user).Collection(u => u.Posts).LoadAsync();
-            
-            var post = user.Posts.FirstOrDefault(post => post.Id == request.Id);
-
-            if (post is null)
-            {
-                throw new BadRequestRestException("Post does not exists");
-            }
-
-            post.Text = request.Text;
-            
-            await _context.SaveChangesAsync();
+            throw new BadRequestRestException("Post does not exists");
         }
 
-        public async Task DeleteAsync(Guid userId, Guid postId)
+        post.Text = request.Text;
+            
+        await _context.SaveChangesAsync();
+    }
+
+    public async Task DeleteAsync(Guid userId, Guid postId)
+    {
+        User user = await _context.Users.FindAsync(userId);
+
+        if (user is null)
         {
-            User user = await _context.Users.FindAsync(userId);
+            throw new BadRequestRestException("User does not exists");
+        }
+            
+        await _context.Entry(user).Collection(u => u.Posts).LoadAsync();
+            
+        var post = user.Posts.FirstOrDefault(post => post.Id == postId);
 
-            if (user is null)
-            {
-                throw new BadRequestRestException("User does not exists");
-            }
-            
-            await _context.Entry(user).Collection(u => u.Posts).LoadAsync();
-            
-            var post = user.Posts.FirstOrDefault(post => post.Id == postId);
-
-            if (post is null)
-            {
-                throw new BadRequestRestException("Post does not exists");
-            }
-
-            if (!user.Posts.Contains(post))
-            {
-                throw new BadRequestRestException("Post does not exists");
-            }
-            
-            _context.Posts.Remove(post);
-            
-            await _context.SaveChangesAsync();
+        if (post is null)
+        {
+            throw new BadRequestRestException("Post does not exists");
         }
 
-        public async Task PublishAsync(Guid postId)
+        if (!user.Posts.Contains(post))
         {
-            Post post = await _context.Posts.FindAsync(postId);
-            
-            if (post is null)
-            {
-                throw new BadRequestRestException("Post does not exists");
-            }
-
-            var publish = new Publish
-            {
-                Post = post
-            };
-
-            await _context.AddAsync(publish);
-
-            await _context.SaveChangesAsync();
+            throw new BadRequestRestException("Post does not exists");
         }
+            
+        _context.Posts.Remove(post);
+            
+        await _context.SaveChangesAsync();
+    }
+
+    public async Task PublishAsync(Guid postId)
+    {
+        Post post = await _context.Posts.FindAsync(postId);
+            
+        if (post is null)
+        {
+            throw new BadRequestRestException("Post does not exists");
+        }
+
+        var publish = new Publish
+        {
+            Post = post
+        };
+
+        await _context.AddAsync(publish);
+
+        await _context.SaveChangesAsync();
     }
 }
