@@ -2,7 +2,6 @@ using System;
 using System.Text;
 using API.Helpers;
 using API.Middlewares;
-using API.Services;
 using API.Services.v1;
 using DAL;
 using DAL.Entities;
@@ -13,37 +12,35 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.OpenApi.Models;
 using FluentValidation.AspNetCore;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ApiExplorer;
-using Microsoft.AspNetCore.Mvc.Versioning;
 using Microsoft.IdentityModel.Tokens;
 
-namespace API
+namespace API;
+
+public class Startup
 {
-    public class Startup
+    public Startup(IConfiguration configuration)
     {
-        public Startup(IConfiguration configuration)
+        Configuration = configuration;
+    }
+
+    public IConfiguration Configuration { get; }
+
+    public void ConfigureServices(IServiceCollection services)
+    {
+        services.AddDbContext<DataContext>(options =>
         {
-            Configuration = configuration;
-        }
+            options.UseSqlServer(Configuration["ConnectionStrings:Default"]);
+        });
 
-        public IConfiguration Configuration { get; }
+        services.AddIdentity<User, IdentityRole<Guid>>()
+            .AddEntityFrameworkStores<DataContext>()
+            .AddDefaultTokenProviders();
 
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddDbContext<DataContext>(options =>
-            {
-                options.UseSqlServer(Configuration["ConnectionStrings:Default"]);
-            });
-
-            services.AddIdentity<User, IdentityRole<Guid>>()
-                    .AddEntityFrameworkStores<DataContext>()
-                    .AddDefaultTokenProviders();
-
-            services.AddAuthentication(options =>
+        services.AddAuthentication(options =>
             {
                 options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
                 options.DefaultScheme = JwtBearerDefaults.AuthenticationScheme;
@@ -65,64 +62,63 @@ namespace API
                 };
             });
             
-            services.AddFluentValidation(options =>
-            {
-                options.DisableDataAnnotationsValidation = false;
-                options.RegisterValidatorsFromAssemblyContaining<Startup>();
-            });
-            
-            // Connect services
-            services.AddTransient<IdentityServiceV1>();
-            services.AddTransient<PostsServiceV1>();
-            services.AddTransient<CommentsServiceV1>();
-            
-            // Connect helpers
-            services.AddTransient<JwtHelper>();
-            
-            services.AddControllers();
-
-            services.AddApiVersioning(options =>
-            {
-                options.DefaultApiVersion = ApiVersion.Default;
-                options.AssumeDefaultVersionWhenUnspecified = true;
-                options.ReportApiVersions = true;
-            });
-            
-            services.AddVersionedApiExplorer(options =>
-            {
-                options.GroupNameFormat = "'v'VVV";
-                options.SubstituteApiVersionInUrl = true;
-            });
-
-            services.AddSwaggerGen();
-            services.ConfigureOptions<ConfigureSwaggerOptions>();
-        }
-
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env, IApiVersionDescriptionProvider provider)
+        services.AddFluentValidation(options =>
         {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-                app.UseSwagger();
-                app.UseSwaggerUI(options =>  
-                {  
-                    foreach (var description in provider.ApiVersionDescriptions)
-                    {
-                        options.SwaggerEndpoint($"/swagger/{description.GroupName}/swagger.json", description.GroupName);  
-                    }  
-                }); 
-            }
+            options.DisableDataAnnotationsValidation = false;
+            options.RegisterValidatorsFromAssemblyContaining<Startup>();
+        });
+            
+        // Connect services
+        services.AddTransient<IdentityServiceV1>();
+        services.AddTransient<PostsServiceV1>();
+        services.AddTransient<CommentsServiceV1>();
+            
+        // Connect helpers
+        services.AddTransient<JwtHelper>();
+            
+        services.AddControllers();
 
-            app.UseMiddleware<ErrorsHandlerMiddleware>();
+        services.AddApiVersioning(options =>
+        {
+            options.DefaultApiVersion = ApiVersion.Default;
+            options.AssumeDefaultVersionWhenUnspecified = true;
+            options.ReportApiVersions = true;
+        });
+            
+        services.AddVersionedApiExplorer(options =>
+        {
+            options.GroupNameFormat = "'v'VVV";
+            options.SubstituteApiVersionInUrl = true;
+        });
 
-            app.UseHttpsRedirection();
+        services.AddSwaggerGen();
+        services.ConfigureOptions<ConfigureSwaggerOptions>();
+    }
 
-            app.UseRouting();
-
-            app.UseAuthentication();
-            app.UseAuthorization();
-
-            app.UseEndpoints(endpoints => { endpoints.MapControllers(); });
+    public void Configure(IApplicationBuilder app, IWebHostEnvironment env, IApiVersionDescriptionProvider provider)
+    {
+        if (env.IsDevelopment())
+        {
+            app.UseDeveloperExceptionPage();
+            app.UseSwagger();
+            app.UseSwaggerUI(options =>  
+            {  
+                foreach (var description in provider.ApiVersionDescriptions)
+                {
+                    options.SwaggerEndpoint($"/swagger/{description.GroupName}/swagger.json", description.GroupName);  
+                }  
+            }); 
         }
+
+        app.UseMiddleware<ErrorsHandlerMiddleware>();
+
+        app.UseHttpsRedirection();
+
+        app.UseRouting();
+
+        app.UseAuthentication();
+        app.UseAuthorization();
+
+        app.UseEndpoints(endpoints => { endpoints.MapControllers(); });
     }
 }
