@@ -5,27 +5,26 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
 
-namespace DAL
-{
-    public class DataContext : IdentityDbContext<User, IdentityRole<Guid>, Guid>
-    {
-        public DataContext(DbContextOptions<DataContext> options)
-            : base(options)
-        {
-        }
+namespace DAL;
 
-        protected override void OnModelCreating(ModelBuilder builder)
-        {
-            builder.ApplyConfiguration(new UserConfiguration());
-            builder.ApplyConfiguration(new PostConfiguration());
-            builder.ApplyConfiguration(new PublishConfiguration());
-            builder.ApplyConfiguration(new CommentConfiguration());
-            
-            base.OnModelCreating(builder);
-        }
-        
-        public DbSet<Post> Posts { get; set; }
-        public DbSet<Publish> Publishes { get; set; }
-        public DbSet<Comment> Comments { get; set; }
+public class DataContext : IdentityDbContext<User, IdentityRole<Guid>, Guid>
+{
+    public DataContext(DbContextOptions<DataContext> options)
+        : base(options)
+    {
     }
+
+    protected override void OnModelCreating(ModelBuilder builder)
+    {
+        builder.ApplyConfiguration(new UserConfiguration());
+        builder.ApplyConfiguration(new PostConfiguration());
+        builder.ApplyConfiguration(new PublishConfiguration());
+        builder.ApplyConfiguration(new CommentConfiguration());
+            
+        base.OnModelCreating(builder);
+    }
+        
+    public DbSet<Post> Posts { get; set; }
+    public DbSet<Publish> Publishes { get; set; }
+    public DbSet<Comment> Comments { get; set; }
 }
