@@ -1,11 +1,10 @@
 ﻿using DAL.Entities.Base;
 
-namespace DAL.Entities
+namespace DAL.Entities;
+
+public class Comment : Entity
 {
-    public class Comment : Entity
-    {
-        public string Text { get; set; }
-        public Post Post { get; set; }
-        public User Author { get; set; }
-    }
+    public string Text { get; set; }
+    public Post Post { get; set; }
+    public User Author { get; set; }
 }
