@@ -1,11 +1,10 @@
 ﻿using System;
 using DAL.Entities.Base;
 
-namespace DAL.Entities
+namespace DAL.Entities;
+
+public class Publish : Entity
 {
-    public class Publish : Entity
-    {
-        public Guid PostId { get; set; }
-        public Post Post { get; set; }
-    }
+    public Guid PostId { get; set; }
+    public Post Post { get; set; }
 }
