@@ -1,24 +1,17 @@
+using API;
 using API.Extensions;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Hosting;
 
-namespace API
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            CreateHostBuilder(args).Build().Seed().Run();
-        }
+var hostBuilder = Host.CreateDefaultBuilder(args);
 
-        public static IHostBuilder CreateHostBuilder(string[] args)
-        {
-            var hostBuilder = Host.CreateDefaultBuilder(args);
-            hostBuilder.ConfigureWebHostDefaults(webBuilder =>
-            {
-                webBuilder.UseStartup<Startup>();
-            });
-            return hostBuilder;
-        }
-    }
-}
+hostBuilder.ConfigureWebHostDefaults(builder =>
+{
+    builder.UseStartup<Startup>();
+});
+
+var host = hostBuilder.Build();
+
+host.Seed();
+
+host.Run();
